Move credentials package
