+++ internal/docker/legacydocker/legacydocker.go
@@ -27,9 +27,9 @@ import (
 	"github.com/rs/zerolog/log"
 
 	"github.com/saucelabs/saucectl/cli/config"
-	"github.com/saucelabs/saucectl/cli/credentials"
 	"github.com/saucelabs/saucectl/cli/streams"
 	"github.com/saucelabs/saucectl/cli/utils"
+	"github.com/saucelabs/saucectl/internal/credentials"
 )
 
 var (
